Move trello-related functions to separate file
